requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-admin-interface==0.28.6
django-allauth==0.61.1
django-colorfield==0.11.0
django-filter==24.1
django-ipware==6.0.4
django-phonenumber-field==7.3.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
phonenumbers==8.13.33
pillow==10.3.0
postmarker==1.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-ipware==2.0.2
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
